task2/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10
django==3.1
djangorestframework==3.11
pytz==2022.1
sqlparse==0.4.2
